Add test duration and success
* Add TestResult
* Refactor test_meta module
* Moved most of generated code into FluvioTest
* First pass at implementing Display for TestResult - Resolves #878
